requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
email-validator
requests
httpx
python-multipart
# Gemini API (newer package - google-genai)
google-generativeai>=0.3.0
# Image Generation (Stable Diffusion with LCM)
diffusers>=0.25.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
Pillow>=10.0.0
jinja2
# Optimization libraries
lcm-solver>=0.1.0
omegaconf
safetensors
python-dotenv
